requirements.txt
pyTelegramBotAPI
python-dotenv
regex
Flask
gunicorn
mysql-connector-python